refactoring of code base
