+++ graphql/schema/index.js
@@ -40,7 +40,7 @@ input EventInput {
 
 type RootQuery {
   events: [Event!]!
-  booking:[Booking!]!
+  bookings:[Booking!]!
 }
 
 type RootMutation {
